Made timeFrameToText more widely accessible
